requirements.txt
yfinance
pandas
numpy
pyarrow
plotly
pandas-ta-openbb
matplotlib
mplfinance
seaborn